SCP-3049 created module structure and version no
